requirements.txt
Django==3.2.4
PyMySQL==1.0.2
django-utils-six==2.0
django-log-to-telegram==2020.1.23
django-axes==5.18.0
cryptography==3.4.7
six==1.16.0
requests==2.25.1
django-environ==0.4.5
pycryptodome